fix(i18n): update translation keys

+++ src/com/gameserver/interfaces/IInventory.java
@@ -11,11 +11,11 @@ public interface IInventory {
     long getVolume();
     long getFreeVolume();
 
-    default boolean canBeStored(long amount, long volume) {
+    default boolean canBeStored(double amount, long volume) {
         return getFreeVolume() > (amount * volume);
     }
 
-    default long getAmountThatCanBeStored(long volume) {
+    default double getAmountThatCanBeStored(long volume) {
         if(volume == 0) return (long) Double.POSITIVE_INFINITY;// TODO: FIX ME
         return Math.max(0, Math.floorDiv(getFreeVolume(), volume));
     }
